feat: added react hook form to datepicker. refactor: added fest files to fest folder in gql
