+++ src/graphql/index.js
@@ -1,12 +1,9 @@
 export * from "./create-discussion";
 export * from "./create-user";
-export * from "./delete-fest";
 export * from "./get-discussions";
 export * from "./get-fest";
 export * from "./get-movies";
 export * from "./get-user-authentication";
 export * from "./get-user-fests";
 export * from "./get-users";
-export * from "./hook/useCurrentUser";
 export * from "./signin-users";
-export * from "./update-fest";
