+++ src/graphql/mutations/fest/delete-fest.js
