+++ src/graphql/mutations/fest/index.js
@@ -1,0 +1,3 @@
+export * from "./create-fest";
+export * from "./delete-fest";
+export * from "./update-fest";
