+++ src/routes/fest-route/fest-route.jsx
@@ -2,7 +2,8 @@ import { useMutation, useQuery } from "@apollo/client";
 import { useEffect, useMemo } from "react";
 import { useParams } from "react-router";
 import { Button, Header, Loading, MovieCardList } from "../../components";
-import { GET_FEST, GET_MOVIES, UPDATE_FEST } from "../../graphql/";
+import { GET_FEST, GET_MOVIES } from "../../graphql/";
+import { UPDATE_FEST } from "../../graphql/mutations/fest";
 import { useModals } from "../../hooks";
 import magGlassDark from "../../images/mag-glass-black.svg";
 import { FestHeader, FestModal, FestSidebar } from "../fest-route";
