+++ src/store/client-context-provider.jsx
@@ -14,7 +14,7 @@ export const ClientContextProvider = ({ children }) => {
         import.meta.env.MODE === "production"
           ? //created environment variables must be prefixed by VITE
             import.meta.env.VITE_API_URI
-          : "https://slopopedia-api-a5fe9aef64e8.herokuapp.com/api/graphql",
+          : "http://localhost:8080/api/graphql",
     });
 
     const authLink = setContext((_, { headers }) => {
@@ -41,7 +41,7 @@ export const ClientContextProvider = ({ children }) => {
           import.meta.env.MODE === "production"
             ? //created environment variables must be prefixed by VITE
               import.meta.env.VITE_API_URI
-            : "https://slopopedia-api-a5fe9aef64e8.herokuapp.com/api/graphql",
+            : "http://localhost:8080/api/graphql",
       });
 
       const authLink = setContext((_, { headers }) => {
